DR-1408 Run hookWrapper#endStep even when the step throws an exception (#39)
- Put hookWrapper().endStep in a finally so that it is always run, even if the step ends by throwing an exception.
- Fix bash script for setting variables by removing whitespace.
Right now, if the step returns a StepResult error, the endStep on a StairwayHook will run. If the step throws an exception however, Stairway will catch the exception and wrap it as a StepResult error, but not run endStep. We should run the StairwayHook#endStep regardless of how the step finishes.