PF-694 Add FlightDebugInfo support for failing after the last step. (#68)
Extend the FlightDebugInfo in Stairway to allow it to be easy to test undoing a flight by setting a failure after the last do step executes.

By adding a field, existing versions of Stairway will not be able to deserialize FlightDebugInfo's serialized by this and newer versions going forward: they will be unable to make a FlightContext. Older versions serializations will be able to be deserialized by this version, as ObjectMapper's FAIL_ON_UNKNOWN_PROPERTIES defaults to false. this is a backwards incompatible change for FlightDebugInfo serialization.
Set DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES to make this not the case in the future.