PF-699 Add FlightDebugInfo support for failing do & undo by Step class name (#72)
Extend Stairway's FlightDebugInfo to allow changing the status of step's by the Step's class name instead of a step index. Step index could be changed by adding more steps to a Flight, but by using the Step.class class name, it's easier to express what's being tested and fail to compile if a test's step name changes.

Allow the setting failures on both do and undo for a Step, as it's important to verify that undo operations are also idempotent to avoid dismal failures.

This will not support the same step class name being run multiple times, but that does not seem to be as useful.