Code reorg: flight context-awareness moved to StairwayThreadPool
Because Stairway flights are submitted to the StairwayThreadPool for execution, it makes more sense to update the threadpool to wrap task submission to handle setting up and clearing flight context for the duration of the flight run on the child thread.
MdcHelper becomes MdcUtils, now a home for less involved utility methods for modifying the MDC.
Expanded tests accordingly: StairwayThreadPool did not previously have targeted unit tests.