Make postgresql a stairway.java-conventions dependency
That way, we set the version in one place, and it will be applied in all projects with how we've set up the plug-in and inheritance.
We discussed that while postgresql is available in the Spring Boot dependency manager, we're only using that for stairctl and our priority is that the postgresql version we're using in stairway project is used everywhere.