StairwayImpl constructor executor init follows existing convention
Mainly because my usage of Optional.ofNullable…orElse… constructs a new default executor every time, and throws it away if the builder has one set.  This is discouraged.

I decided to adhere with the existing convention for initializing instance variables from the StairwayBuilder object, which makes the constructor easier to follow.  If it's updated in the future with Optional handling, that would be acceptable as long as it's updated completely.