[DCJ-457] Stairway can be built with a provided ThreadPoolTaskExecutor (#149)
* [DCJ-457] Stairway can be built with a provided ThreadPoolTaskExecutor

This allows calling services to bring their own instrumented executor to serve as their Stairway's threadpool.

If unspecified in StairwayBuilder, a DefaultThreadPoolTaskExecutor will be constructed and used.

DefaultThreadPoolTaskExecutor is a public class: calling services may elect to instantiate it themselves (e.g. a Spring Boot service would register this as a Bean for automatic actuator instrumentation).

* Also assert that DefaultThreadPoolTaskExecutors are running

* Docstring references default maxParallelFlights directly

If this value were to change in the future, the javadoc rendered view would reflect the change.

* StairwayImpl constructor executor init follows existing convention

Mainly because my usage of Optional.ofNullable…orElse… constructs a new default executor every time, and throws it away if the builder has one set.  This is discouraged.

I decided to adhere with the existing convention for initializing instance variables from the StairwayBuilder object, which makes the constructor easier to follow.  If it's updated in the future with Optional handling, that would be acceptable as long as it's updated completely.

* Update mavenLocal development docs

This matches TDR's current build file -- the useMavenLocal flag makes it a little easier to leverage locally published artifacts on demand.