[DCJ-507] Flights queued in GCP Pub-Sub include calling thread context
Context-aware Stairway previously assumed that the calling thread would have mapped diagnostic context (MDC) to persist to the child threads spawned for flight execution.

This was not the case when Stairway is configured with a work queue enabled (e.g. GCP Pub-Sub).  Here, flight information was written to the pub-sub topic without the calling thread's context, so when a message was processed and deserialized any logs emitted from that flight would be missing context set by the original calling thread (e.g. request ID).

- Expanded QueueMessageReady to store calling thread context on construction
- QueueMessageReady.process sets the MDC using its stored calling thread context, which then makes the flight context-aware
- Expanded unit tests